splt test main into multiple files that are route specific
